src/requirements.txt
transformers
torchaudio
sentencepiece
pydub
openai-whisper